Modify migration

+++ src/resources/views/migration.blade.php
@@ -15,17 +15,33 @@ class {{$className}} extends Migration
     public function up()
     {
 
+@if(!$isModify)
 Schema::create('{{$tableName}}', function(Blueprint $table) {
-$table->id();@foreach($list as $ind=> $data)
+$table->id();
+$table->timestamps();
+@else
+    Schema::table('{{$tableName}}', function(Blueprint $table) {
+@endif
+
+@foreach($list as $ind=> $data)
 
     $table @foreach( $data as $fun=> $arg)  ->{{$fun}}({{$arg}}) @endforeach;@endforeach
 
-$table->timestamps();
 });
 
-
+}
     public function down()
     {
-        Schema::dropIfExists('{{$tableName}}');
+@if(!$isModify)
+    Schema::dropIfExists('{{$tableName}}');
+@else
+    Schema::table('{{$tableName}}', function (Blueprint $table) {
+         $table->dropColumn([
+    @foreach($list as $ind=> $data)
+    '{{$ind}}',
+    @endforeach
+    ]);
+    });
+@endif
     }
 }
